Save and Load Topology View (#46)
A Topology Viewer page (plugins/nextbox-ui/topology/) now allows you to save a current layout to the database.
A saved topology is a "snapshot" of the topology state on the screen with all associated attributes.
A saved topology can then be loaded from the Topology Viewer page. Overall device layout and their relative coordinates will be preserved. General plugin Hide/Display controls are also preserved on the view save/load.
Device and link changes taken between save and load actions are not currently considered. Load Saved View shows a topology state in the past as is.
Tracking changes is a matter of further development.

* Save Topology View feature initial commit
Add general workflow for current layout saving.
Loading from the created snapshot is not yet implemented.

* Implement Load Saved View
It is possible to load a saved topology from on a Topology Viewer page.